complete

+++ src/main/java/com/board/basic/article/ArticleController.java
@@ -25,8 +25,8 @@ public class ArticleController {
     private final UserService userService;
 
     @GetMapping("/list")
-    public String list(Model model) {
-        List<Article> articleList = this.articleService.getList();
+    public String list(Model model, @RequestParam(value = "keyword", defaultValue = "") String keyword) {
+        List<Article> articleList = this.articleService.getList(keyword);
         model.addAttribute("articleList", articleList);
         return "article_list";
     }
@@ -88,9 +88,15 @@ public class ArticleController {
 
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable("id") Integer id, Model model) {
+    public String delete(@PathVariable("id") Integer id, Model model, Principal principal) {
         Article article = this.articleService.getArticle(id);
+
         model.addAttribute("article", article);
+
+        if (!article.getUser().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제 권한이 없습니다.");
+        }
+
         this.articleService.delete(article);
         return "redirect:/article/list";
     }
